move test plugin into test folder
