+++ test/testPlugin/tasks/pluginTask.js
@@ -1,4 +1,4 @@
-const ActionHero = require('./../../index.js')
+const ActionHero = require('./../../../index.js')
 
 module.exports = class PlutingTask extends ActionHero.Task {
   constructor () {
